move initial chart installation
