+++ integration/setup/setup.go
@@ -3,19 +3,13 @@
 package setup
 
 import (
-	"fmt"
 	"log"
-	"net/http"
 	"os"
-	"strconv"
 	"testing"
-	"time"
 
-	"github.com/cenkalti/backoff"
 	"github.com/giantswarm/apprclient"
 	"github.com/giantswarm/e2e-harness/pkg/framework"
 	"github.com/giantswarm/helmclient"
-	"github.com/giantswarm/k8sportforward"
 	"github.com/giantswarm/microerror"
 	"github.com/giantswarm/micrologger"
 	"github.com/spf13/afero"
@@ -82,11 +76,6 @@ func initializeCNR(f *framework.Host, helmClient *helmclient.Client) error {
 		return microerror.Mask(err)
 	}
 
-	err = installInitialCharts(f)
-	if err != nil {
-		return microerror.Mask(err)
-	}
-
 	return nil
 }
 
@@ -123,119 +112,4 @@ func installCNR(f *framework.Host, helmClient *helmclient.Client) error {
 	}
 
 	return nil
-}
-
-func installInitialCharts(f *framework.Host) error {
-	fwc := k8sportforward.Config{
-		RestConfig: f.RestConfig(),
-	}
-
-	fw, err := k8sportforward.New(fwc)
-	if err != nil {
-		return microerror.Mask(err)
-	}
-
-	podName, err := waitForPod(f, "giantswarm", "app=cnr-server")
-	if err != nil {
-		return microerror.Mask(err)
-	}
-	tc := k8sportforward.TunnelConfig{
-		Remote:    5000,
-		Namespace: "giantswarm",
-		PodName:   podName,
-	}
-	tunnel, err := fw.ForwardPort(tc)
-	if err != nil {
-		return microerror.Mask(err)
-	}
-
-	serverAddress := "http://localhost:" + strconv.Itoa(tunnel.Local)
-	err = waitForServer(f, serverAddress+"/cnr/api/v1/packages")
-	if err != nil {
-		return microerror.Mask(fmt.Errorf("server didn't come up on time"))
-	}
-
-	l, err := micrologger.New(micrologger.Config{})
-	if err != nil {
-		return microerror.Mask(err)
-	}
-
-	c := apprclient.Config{
-		Fs:     afero.NewOsFs(),
-		Logger: l,
-
-		Address:      serverAddress,
-		Organization: "giantswarm",
-	}
-
-	a, err := apprclient.New(c)
-	if err != nil {
-		return microerror.Mask(err)
-	}
-
-	err = a.PushChartTarball("tb-chart", "5.5.5", "/e2e/fixtures/tb-chart-5.5.5.tgz")
-	if err != nil {
-		return microerror.Mask(err)
-	}
-
-	err = a.PromoteChart("tb-chart", "5.5.5", "5-5-beta")
-	if err != nil {
-		return microerror.Mask(err)
-	}
-
-	err = a.PushChartTarball("tb-chart", "5.6.0", "/e2e/fixtures/tb-chart-5.6.0.tgz")
-	if err != nil {
-		return microerror.Mask(err)
-	}
-
-	err = a.PromoteChart("tb-chart", "5.6.0", "5-6-beta")
-	if err != nil {
-		return microerror.Mask(err)
-	}
-
-	return nil
-}
-
-func waitForServer(f *framework.Host, url string) error {
-	var err error
-
-	operation := func() error {
-		_, err := http.Get(url)
-		if err != nil {
-			return fmt.Errorf("could not retrieve %s: %v", url, err)
-		}
-		return nil
-	}
-
-	notify := func(err error, t time.Duration) {
-		log.Printf("waiting for server at %s: %v", t, err)
-	}
-
-	err = backoff.RetryNotify(operation, backoff.NewExponentialBackOff(), notify)
-	if err != nil {
-		return microerror.Mask(err)
-	}
-	return nil
-}
-
-func waitForPod(f *framework.Host, ns, selector string) (string, error) {
-	var err error
-	var podName string
-	operation := func() error {
-		podName, err = f.GetPodName(ns, selector)
-		if err != nil {
-			return fmt.Errorf("could not retrieve pod %q on %q: %v", selector, ns, err)
-		}
-		return nil
-	}
-
-	notify := func(err error, t time.Duration) {
-		log.Printf("waiting for pod at %s: %v", t, err)
-	}
-
-	err = backoff.RetryNotify(operation, backoff.NewExponentialBackOff(), notify)
-	if err != nil {
-		return "", microerror.Mask(err)
-	}
-	return podName, nil
 }
